Added Thread registering feature
Modified Thread test accordingly
TODO: implement Message dispatching
TODO: add a Thread to the Cerberus Provider Core
TODO: make the core log to file independently

+++ src/message/messagetemplate.cpp
@@ -31,7 +31,7 @@ std::string MessageTemplate::name() const
     return m_name;
 }
 //=============================================================================
-uint32_t MessageTemplate::count() const
+size_t MessageTemplate::count() const
 {
     return m_types.size();
 }
